requirements.txt.txt
rustplus==5.6.8
asyncio==3.4.3
discord.py